requirements.txt
openai-whisper>=20231117
torch>=2.0.0
tqdm>=4.65.0
audible>=0.8.2
